requirements.txt
# Core dependencies
pydantic-settings==2.10.1
PyGithub==2.7.0
langchain-core==0.3.75
torch>=2.0.0
python-dotenv==1.1.1

# Optional dependencies for future features
# fastapi>=0.100.0  # For webhook server (Phase 4)
# apscheduler>=3.10.0  # For scheduling (Phase 4)
# requests>=2.31.0  # For Swimlane API (Phase 5)

